claude-agent-sdk-in-action/code-examples/chapter-09/requirements.txt
# 第 9 章：多層次協調與元 Agent
# Meta Agent 系統依賴套件

# Claude Agent SDK
anthropic==0.39.0

# 非同步支援
asyncio-mqtt==0.16.1

# 資料處理
pydantic==2.5.0
python-dotenv==1.0.0

# 進度視覺化
rich==13.7.0

# 系統監控
psutil==5.9.6

# 日誌
structlog==24.1.0

# 測試
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# 型別檢查
mypy==1.7.1

# 程式碼品質
black==23.12.0
flake8==6.1.0
isort==5.13.0
